requirements.txt
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
Pillow==10.4.0
gunicorn==22.0.0
pytest==8.3.3
pytest-django==4.9.0